feat: Add useCardInfo custom hook
