+++ src/pages/card-add/card-form/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useCardInfo } from './useCardInfo'
